requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
numpy==1.24.3
pandas==2.0.3
tensorflow==2.13.0
tensorflow-hub==0.14.0
PyPDF2==3.0.1
python-docx==0.8.11
gunicorn==20.1.0
reportlab==3.6.12
requests==2.28.1 